+++ app/Http/Controllers/Admin/DataManagement/VenuesController.php
@@ -43,7 +43,7 @@ class VenuesController extends Controller
      */
     public function store(Request $request)
     {
-        
+
         Venue::create($request->all());
 
         Flash::success('Venue added!');
@@ -54,7 +54,7 @@ class VenuesController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      *
      * @return Response
      */
@@ -68,7 +68,7 @@ class VenuesController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      *
      * @return Response
      */
@@ -82,13 +82,13 @@ class VenuesController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  int  $id
+     * @param  int $id
      *
      * @return Response
      */
     public function update($id, Request $request)
     {
-        
+
         $venue = Venue::findOrFail($id);
         $venue->update($request->all());
 
@@ -100,16 +100,20 @@ class VenuesController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      *
      * @return Response
      */
     public function destroy($id)
     {
-        Venue::destroy($id);
-
-        Flash::success('Venue deleted!');
-
+        $canBeDeleted = empty(Venue::find($id)->fixtures->toArray());
+        if ($canBeDeleted) {
+            Venue::destroy($id);
+            Flash::success('Venue deleted!');
+        } else {
+            Flash::error('Cannot delete because they are existing fixtures at this venue.');
+        }
+        
         return redirect('admin/data-management/venues');
     }
